[client] rename python client dir to py_client
